+++ payment/payment-service/src/main/java/shop/woowasap/payment/service/PaymentService.java
@@ -2,8 +2,6 @@ package shop.woowasap.payment.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.woowasap.core.id.api.IdGenerator;
@@ -12,11 +10,8 @@ import shop.woowasap.order.domain.Order;
 import shop.woowasap.order.domain.exception.DoesNotFindOrderException;
 import shop.woowasap.order.domain.in.OrderConnector;
 import shop.woowasap.order.domain.in.event.PaySuccessEvent;
-import shop.woowasap.order.domain.out.event.StockFailEvent;
-import shop.woowasap.order.domain.out.event.StockSuccessEvent;
 import shop.woowasap.payment.domain.PayStatus;
 import shop.woowasap.payment.domain.Payment;
-import shop.woowasap.payment.domain.exception.DoesNotFindPaymentException;
 import shop.woowasap.payment.domain.exception.PayUserNotMatchException;
 import shop.woowasap.payment.domain.in.PaymentUseCase;
 import shop.woowasap.payment.domain.in.request.PaymentRequest;
@@ -24,8 +19,8 @@ import shop.woowasap.payment.domain.in.response.PaymentResponse;
 import shop.woowasap.payment.domain.out.PaymentRepository;
 
 @Service
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class PaymentService implements PaymentUseCase {
 
     private final ApplicationEventPublisher applicationEventPublisher;
@@ -72,23 +67,5 @@ public class PaymentService implements PaymentUseCase {
             .purchasedMoney(order.getTotalPrice())
             .createdAt(timeUtil.now())
             .build();
-    }
-
-    @Async
-    @Transactional
-    @EventListener(StockSuccessEvent.class)
-    public void successPayment(final StockSuccessEvent stockSuccessEvent) {
-        final Payment payment = paymentRepository.findByOrderId(stockSuccessEvent.orderId())
-            .orElseThrow(() -> new DoesNotFindPaymentException(stockSuccessEvent.orderId()));
-        paymentRepository.save(payment.changeStatus(PayStatus.SUCCESS));
-    }
-
-    @Async
-    @Transactional
-    @EventListener(StockFailEvent.class)
-    public void cancelPayment(final StockFailEvent stockFailEvent) {
-        final Payment payment = paymentRepository.findByOrderId(stockFailEvent.orderId())
-            .orElseThrow(() -> new DoesNotFindPaymentException(stockFailEvent.orderId()));
-        paymentRepository.save(payment.changeStatus(PayStatus.CANCELD));
     }
 }
